Fixed spec and changed 'array' to 'list'. Changed the fetching of 'val' to use Erlang's array.get, but 3 of the tests do not pass.
